gtk-doc.make: move similar rules together
This patch moves several definitions around so that it's possible to
group rules for same DOC_MODULE together.

Besides, this patch also eliminates the use of HTML_BUILD_STAMP and
PDF_BUILD_STAMP. They're barely helpful, and by replacing them with
hard coded names, it will be easier to convert that Makefile to
support subdir-objects.

Signed-off-by: You-Sheng Yang <vicamo@gmail.com>