gtk-doc.make: add prefix to module variables
In order to support subdir-objects style Makefile, gtk-doc.make
should be able to handle multiple doc module definitions, which
follows there must be a way to separate variables of different
modules. In this patch we turn most lines into Makefile funtions and
therefore gain the ability to evaluate them with a prefix parameter.

Signed-off-by: You-Sheng Yang <vicamo@gmail.com>